rename: `ChannelUserStatus`를 `@pages`에서 `@layouts`로 이동 #246
